requirements.txt
Django==5.0
djangorestframework==3.14.0
pygame==2.5.2
Pillow==10.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.1 